Move app.js code to separate component
